// src/bots/tiktok-videos/callback-query.enum.ts
export enum TikTokVideosBotCallbackQuery {
  Go = 'go',

  // Editor
  OpenEditor = 'open_editor',
  CloseEditor = 'editor_close',
  ProcessVideo = 'editor_process_video',
  SoundSettings = 'editor_sound_settings',
  RatioSettings = 'editor_ratio_settings',
  RollManagement = 'editor_roll_management',
  SelectTemplate = 'editor_select_template',
  BackdropSettings = 'editor_backdrop_selection',
  SelectVerticalPosition = 'editor_vertical_position',
}